# radix: optimize for our use case