--- a/pkg/radix/tree_test.go
+++ b/pkg/radix/tree_test.go
@@ -1,12 +1,26 @@
 package radix
 
 import (
-	crand "crypto/rand"
+	"crypto/rand"
 	"fmt"
 	"reflect"
-	"sort"
 	"testing"
 )
+
+// generateUUID is used to generate a random UUID
+func generateUUID() string {
+	buf := make([]byte, 16)
+	if _, err := rand.Read(buf); err != nil {
+		panic(fmt.Errorf("failed to read random bytes: %v", err))
+	}
+
+	return fmt.Sprintf("%08x-%04x-%04x-%04x-%12x",
+		buf[0:4],
+		buf[4:6],
+		buf[6:8],
+		buf[8:10],
+		buf[10:16])
+}
 
 func TestRadix(t *testing.T) {
 	var min, max string
@@ -27,32 +41,18 @@ func TestRadix(t *testing.T) {
 		t.Fatalf("bad length: %v %v", r.Len(), len(inp))
 	}
 
-	r.Walk(func(k string, v int) bool {
-		return false
-	})
-
-	for k, v := range inp {
-		out, ok := r.Get(k)
-		if !ok {
-			t.Fatalf("missing key: %v", k)
-		}
-		if out != v {
-			t.Fatalf("value mis-match: %v %v", out, v)
-		}
-	}
-
 	// Check min and max
 	outMin, _, _ := r.Minimum()
-	if outMin != min {
-		t.Fatalf("bad minimum: %v %v", outMin, min)
+	if string(outMin) != min {
+		t.Fatalf("bad minimum: %s %v", outMin, min)
 	}
 	outMax, _, _ := r.Maximum()
-	if outMax != max {
-		t.Fatalf("bad maximum: %v %v", outMax, max)
+	if string(outMax) != max {
+		t.Fatalf("bad maximum: %s %v", outMax, max)
 	}
 
 	for k, v := range inp {
-		out, ok := r.Delete(k)
+		out, ok := r.Get([]byte(k))
 		if !ok {
 			t.Fatalf("missing key: %v", k)
 		}
@@ -60,47 +60,7 @@ func TestRadix(t *testing.T) {
 			t.Fatalf("value mis-match: %v %v", out, v)
 		}
 	}
-	if r.Len() != 0 {
-		t.Fatalf("bad length: %v", r.Len())
-	}
-}
 
-func TestRoot(t *testing.T) {
-	r := New()
-	_, ok := r.Delete("")
-	if ok {
-		t.Fatalf("bad")
-	}
-	_, ok = r.Insert("", 1)
-	if !ok {
-		t.Fatalf("bad")
-	}
-	val, ok := r.Get("")
-	if !ok || val != 1 {
-		t.Fatalf("bad: %v", val)
-	}
-	val, ok = r.Delete("")
-	if !ok || val != 1 {
-		t.Fatalf("bad: %v", val)
-	}
-}
-
-func TestDelete(t *testing.T) {
-
-	r := New()
-
-	s := []string{"", "A", "AB"}
-
-	for _, ss := range s {
-		r.Insert(ss, 1)
-	}
-
-	for _, ss := range s {
-		_, ok := r.Delete(ss)
-		if !ok {
-			t.Fatalf("bad %q", ss)
-		}
-	}
 }
 
 func TestDeletePrefix(t *testing.T) {
@@ -122,220 +82,21 @@ func TestDeletePrefix(t *testing.T) {
 	for _, test := range cases {
 		r := New()
 		for _, ss := range test.inp {
-			r.Insert(ss, 1)
+			r.Insert([]byte(ss), 1)
 		}
 
-		deleted := r.DeletePrefix(test.prefix)
+		deleted := r.DeletePrefix([]byte(test.prefix))
 		if deleted != test.numDeleted {
 			t.Fatalf("Bad delete, expected %v to be deleted but got %v", test.numDeleted, deleted)
 		}
 
 		out := []string{}
-		fn := func(s string, v int) bool {
-			out = append(out, s)
+		fn := func(s []byte, v int) bool {
+			out = append(out, string(s))
 			return false
 		}
-		r.Walk(fn)
+		recursiveWalk(r.root, fn)
 
-		if !reflect.DeepEqual(out, test.out) {
-			t.Fatalf("mis-match: %v %v", out, test.out)
-		}
-	}
-}
-
-func TestLongestPrefix(t *testing.T) {
-	r := New()
-
-	keys := []string{
-		"",
-		"foo",
-		"foobar",
-		"foobarbaz",
-		"foobarbazzip",
-		"foozip",
-	}
-	for _, k := range keys {
-		r.Insert(k, 1)
-	}
-	if r.Len() != len(keys) {
-		t.Fatalf("bad len: %v %v", r.Len(), len(keys))
-	}
-
-	type exp struct {
-		inp string
-		out string
-	}
-	cases := []exp{
-		{"a", ""},
-		{"abc", ""},
-		{"fo", ""},
-		{"foo", "foo"},
-		{"foob", "foo"},
-		{"foobar", "foobar"},
-		{"foobarba", "foobar"},
-		{"foobarbaz", "foobarbaz"},
-		{"foobarbazzi", "foobarbaz"},
-		{"foobarbazzip", "foobarbazzip"},
-		{"foozi", "foo"},
-		{"foozip", "foozip"},
-		{"foozipzap", "foozip"},
-	}
-	for _, test := range cases {
-		m, _, ok := r.LongestPrefix(test.inp)
-		if !ok {
-			t.Fatalf("no match: %v", test)
-		}
-		if m != test.out {
-			t.Fatalf("mis-match: %v %v", m, test)
-		}
-	}
-}
-
-func TestWalkPrefix(t *testing.T) {
-	r := New()
-
-	keys := []string{
-		"foobar",
-		"foo/bar/baz",
-		"foo/baz/bar",
-		"foo/zip/zap",
-		"zipzap",
-	}
-	for _, k := range keys {
-		r.Insert(k, 1)
-	}
-	if r.Len() != len(keys) {
-		t.Fatalf("bad len: %v %v", r.Len(), len(keys))
-	}
-
-	type exp struct {
-		inp string
-		out []string
-	}
-	cases := []exp{
-		{
-			"f",
-			[]string{"foobar", "foo/bar/baz", "foo/baz/bar", "foo/zip/zap"},
-		},
-		{
-			"foo",
-			[]string{"foobar", "foo/bar/baz", "foo/baz/bar", "foo/zip/zap"},
-		},
-		{
-			"foob",
-			[]string{"foobar"},
-		},
-		{
-			"foo/",
-			[]string{"foo/bar/baz", "foo/baz/bar", "foo/zip/zap"},
-		},
-		{
-			"foo/b",
-			[]string{"foo/bar/baz", "foo/baz/bar"},
-		},
-		{
-			"foo/ba",
-			[]string{"foo/bar/baz", "foo/baz/bar"},
-		},
-		{
-			"foo/bar",
-			[]string{"foo/bar/baz"},
-		},
-		{
-			"foo/bar/baz",
-			[]string{"foo/bar/baz"},
-		},
-		{
-			"foo/bar/bazoo",
-			[]string{},
-		},
-		{
-			"z",
-			[]string{"zipzap"},
-		},
-	}
-
-	for _, test := range cases {
-		out := []string{}
-		fn := func(s string, v int) bool {
-			out = append(out, s)
-			return false
-		}
-		r.WalkPrefix(test.inp, fn)
-		sort.Strings(out)
-		sort.Strings(test.out)
-		if !reflect.DeepEqual(out, test.out) {
-			t.Fatalf("mis-match: %v %v", out, test.out)
-		}
-	}
-}
-
-func TestWalkPath(t *testing.T) {
-	r := New()
-
-	keys := []string{
-		"foo",
-		"foo/bar",
-		"foo/bar/baz",
-		"foo/baz/bar",
-		"foo/zip/zap",
-		"zipzap",
-	}
-	for _, k := range keys {
-		r.Insert(k, 1)
-	}
-	if r.Len() != len(keys) {
-		t.Fatalf("bad len: %v %v", r.Len(), len(keys))
-	}
-
-	type exp struct {
-		inp string
-		out []string
-	}
-	cases := []exp{
-		{
-			"f",
-			[]string{},
-		},
-		{
-			"foo",
-			[]string{"foo"},
-		},
-		{
-			"foo/",
-			[]string{"foo"},
-		},
-		{
-			"foo/ba",
-			[]string{"foo"},
-		},
-		{
-			"foo/bar",
-			[]string{"foo", "foo/bar"},
-		},
-		{
-			"foo/bar/baz",
-			[]string{"foo", "foo/bar", "foo/bar/baz"},
-		},
-		{
-			"foo/bar/bazoo",
-			[]string{"foo", "foo/bar", "foo/bar/baz"},
-		},
-		{
-			"z",
-			[]string{},
-		},
-	}
-
-	for _, test := range cases {
-		out := []string{}
-		fn := func(s string, v int) bool {
-			out = append(out, s)
-			return false
-		}
-		r.WalkPath(test.inp, fn)
-		sort.Strings(out)
-		sort.Strings(test.out)
 		if !reflect.DeepEqual(out, test.out) {
 			t.Fatalf("mis-match: %v %v", out, test.out)
 		}
@@ -344,7 +105,7 @@ func TestWalkPath(t *testing.T) {
 
 func TestInsert_Duplicate(t *testing.T) {
 	r := New()
-	vv, ok := r.Insert("cpu", 1)
+	vv, ok := r.Insert([]byte("cpu"), 1)
 	if vv != 1 {
 		t.Fatalf("value mismatch: got %v, exp %v", vv, 1)
 	}
@@ -354,7 +115,7 @@ func TestInsert_Duplicate(t *testing.T) {
 	}
 
 	// Insert a dup with a different type should fail
-	vv, ok = r.Insert("cpu", 2)
+	vv, ok = r.Insert([]byte("cpu"), 2)
 	if vv != 1 {
 		t.Fatalf("value mismatch: got %v, exp %v", vv, 1)
 	}
@@ -364,33 +125,24 @@ func TestInsert_Duplicate(t *testing.T) {
 	}
 }
 
-// generateUUID is used to generate a random UUID
-func generateUUID() string {
-	buf := make([]byte, 16)
-	if _, err := crand.Read(buf); err != nil {
-		panic(fmt.Errorf("failed to read random bytes: %v", err))
-	}
-
-	return fmt.Sprintf("%08x-%04x-%04x-%04x-%12x",
-		buf[0:4],
-		buf[4:6],
-		buf[6:8],
-		buf[8:10],
-		buf[10:16])
-}
+//
+// benchmarks
+//
 
 func BenchmarkTree_Insert(b *testing.B) {
 	t := New()
 
-	keys := make([]string, 0, 10000)
+	keys := make([][]byte, 0, 10000)
 	for i := 0; i < cap(keys); i++ {
-		k := fmt.Sprintf("cpu,host=%d", i)
+		k := []byte(fmt.Sprintf("cpu,host=%d", i))
 		if v, ok := t.Insert(k, 1); v != 1 || !ok {
 			b.Fatalf("insert failed: %v != 1 || !%v", v, ok)
 		}
 		keys = append(keys, k)
 	}
 
+	b.SetBytes(int64(len(keys)))
+	b.ReportAllocs()
 	b.ResetTimer()
 
 	for j := 0; j < b.N; j++ {
@@ -398,6 +150,25 @@ func BenchmarkTree_Insert(b *testing.B) {
 			if v, ok := t.Insert(key, 1); v != 1 || ok {
 				b.Fatalf("insert failed: %v != 1 || !%v", v, ok)
 			}
+		}
+	}
+}
+
+func BenchmarkTree_InsertNew(b *testing.B) {
+	keys := make([][]byte, 0, 10000)
+	for i := 0; i < cap(keys); i++ {
+		k := []byte(fmt.Sprintf("cpu,host=%d", i))
+		keys = append(keys, k)
+	}
+
+	b.SetBytes(int64(len(keys)))
+	b.ReportAllocs()
+	b.ResetTimer()
+
+	for j := 0; j < b.N; j++ {
+		t := New()
+		for _, key := range keys {
+			t.Insert(key, 1)
 		}
 	}
 }