@hercules-influx 2018-05-10
During a run of [megacheck](https://github.com/dominikh/go-tools) the following issues were discovered:
```/tmp/037151180/src/github.com/influxdata/influxdb/cmd/influx-tools/importer/importer.go:73:3: this value of rpi is never used (SA4006)
```